requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
azure-monitor-opentelemetry==1.1.1
pydantic==2.5.0
gunicorn==21.2.0